Removed two-argument interval constructor
This avoids confusion about whether b is the interval size,
the last item, or one past the last item in the Interval(a,b)
constructor.

If b is the size, use Interval::baseSize(a, b)

If b is the last item, use Interval::hull(a, b)

If b is one past the end use Interval::baseSize(a, b-a) or
Interval::hull(a, b-1).