Commented out the specializations for boost::property_traits
I'm not sure how to specialize these in a way that matches for
any Sawyer::Boost::ConstVertexIdPropertyMap<>, and the specializations
as written are wrong and cause problems when Sawyer graphs are
mixed with other graph implementations.

The changes in graphBoost.C demonstrate how users can work around
this, namely by instantiating a particular ConstVertexIdPropertyMap<>
type and then specializing boost::property_traits for that type.