Removed the acyclic check for the thread workers
The ThreadWorkers now does as much work as possible and then reports
cycles at the very end. We do it this way because detecting cycles
ends up being a side effect of scheduling work and is thus faster. If
users want to check for cycles before doing any work they can still
do that by calling the predicate explicitly.