Fix subgraph isomorphism on macOS
On macOS, we can't construct an std::vector<size_t>::const_iterator
from a size_t* because the constructor is private.  Therefore, we
pretend that size_t* is an iterator since we're not going to be
using it for anything except its operator*() const, operator++(),
and operator==(const size_t*) const.