Switched from STL to boost::iterator::iterator_facade
In preparation for fixing fixing BOOST_REVERSE_FOREACH when compiled
with GCC optimizations that think "*iter" returns a const reference
which is valid even after the next "++iter". In actualitiy, since
this is a generator rather than an iteration over existing values,
referenced value changes each time ++iter is called.